# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Nutrition tracking API.\n\nError code catalog:\n- See docs/openapi-error-codes.md in repository root.",
        "title": "Nutrition API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/daily-totals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Get daily nutrition totals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DailyTotalsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/meals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "List meals by user and date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.MealResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Create meal",
                "parameters": [
                    {
                        "description": "Meal payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMealRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.MealResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/meals/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Get meal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MealResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "meals"
                ],
                "summary": "Delete meal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Update meal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Meal update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMealRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MealResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/meals/{id}/items": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Add meal item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Meal item payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddMealItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.MealItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/meals/{meal_id}/items/{item_id}": {
            "delete": {
                "tags": [
                    "meals"
                ],
                "summary": "Delete meal item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "meal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Meal Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Update meal item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "meal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Meal Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Meal item update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMealItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MealItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddMealItemRequest": {
            "type": "object",
            "properties": {
                "food_id": {
                    "description": "Food source ID (mutually exclusive with recipe_id).",
                    "type": "integer",
                    "example": 1
                },
                "recipe_id": {
                    "description": "Recipe source ID (mutually exclusive with food_id).",
                    "type": "integer",
                    "example": 1
                },
                "weight_g": {
                    "description": "Item consumed weight in grams.",
                    "type": "number",
                    "example": 150
                }
            }
        },
...
        "dto.CreateMealRequest": {
            "type": "object",
            "properties": {
                "eaten_at": {
                    "description": "Meal time in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "items": {
                    "description": "Optional items created atomically with meal.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AddMealItemRequest"
                    }
                },
                "meal_type": {
                    "description": "Meal type enum.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/meal.MealType"
                        }
                    ],
                    "example": "lunch"
                }
            }
        },
...
        "dto.UpdateMealRequest": {
            "type": "object",
            "properties": {
                "eaten_at": {
                    "description": "Optional meal time in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T19:00:00Z"
                },
                "meal_type": {
                    "description": "Optional meal type enum.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/meal.MealType"
                        }
                    ],
                    "example": "dinner"
                }
            }
        },
...
        "handlers.DailyTotalsResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Target date in YYYY-MM-DD.",
                    "type": "string",
                    "example": "2026-02-17"
                },
                "total_carbs_g": {
                    "description": "Aggregated carbohydrate grams for the day.",
                    "type": "number",
                    "example": 220
                },
                "total_fat_g": {
                    "description": "Aggregated fat grams for the day.",
                    "type": "number",
                    "example": 70
                },
                "total_kcal": {
                    "description": "Aggregated kcal for the day.",
                    "type": "number",
                    "example": 2100
                },
                "total_protein_g": {
                    "description": "Aggregated protein grams for the day.",
                    "type": "number",
                    "example": 140
                }
            }
        },
...
        "handlers.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.APIError"
                }
            }
        },
...
        "handlers.MealResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Creation timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "eaten_at": {
                    "description": "Meal timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "id": {
                    "description": "Meal ID.",
                    "type": "integer",
                    "example": 1
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.MealItemResponse"
                    }
                },
                "meal_type": {
                    "description": "Meal type.",
                    "type": "string",
                    "example": "lunch"
                },
                "total_carbs_g": {
                    "description": "Aggregated carbohydrate grams for this meal.",
                    "type": "number",
                    "example": 75
                },
                "total_fat_g": {
                    "description": "Aggregated fat grams for this meal.",
                    "type": "number",
                    "example": 2.1
                },
                "total_kcal": {
                    "description": "Aggregated kcal for this meal.",
                    "type": "number",
                    "example": 350
                },
                "total_protein_g": {
                    "description": "Aggregated protein grams for this meal.",
                    "type": "number",
                    "example": 10.5
                },
                "updated_at": {
                    "description": "Last update timestamp in RFC3339 UTC.",
                    "type": "string",
                    "example": "2026-02-17T12:00:00Z"
                },
                "user_id": {
                    "description": "Owner user ID.",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}